+++ opusfilter/util.py
@@ -3,6 +3,7 @@
 import bz2
 import gzip
 import io
+import itertools
 import logging
 import lzma
 import os
@@ -11,8 +12,83 @@ import requests
 from tqdm import tqdm
 import ruamel.yaml
 
+from . import ConfigurationError
+
 
 logger = logging.getLogger(__name__)
+
+
+class FakeList:
+    """Implements __getitem__ that returns always the same value"""
+
+    def __init__(self, value):
+        self.value = value
+
+    def __getitem__(self, key):
+        return self.value
+
+
+def check_args_compability(*args, required_types=None, choices=None, names=None):
+    """Check that arguments of single value or list of values are compatible
+
+    - Nth argument (plain value or within a list) must be instance of required_types[N] (if defined)
+    - Nth argument (plain value or within a list) must be in of choices[N] (if defined)
+    - If some arguments are lists, they have to be of the same length
+    - If some arguments are lists, any non-list values are expanded to lists of the same length
+    - If none of the arguments are lists, arguments are expanded to FakeList objects
+
+    Return the input arguments expanded to lists or FakeLists when needed.
+
+    """
+
+    def type_error_msg(idx, type_, value):
+        name = names[idx] if names else str(idx + 1)
+        return f"Values of argument '{name}' are not of the type {type_.__name__}: {value}"
+
+    def value_error_msg(idx, choices, value):
+        name = names[idx] if names else str(idx + 1)
+        return f"Values of argument '{name}' are not one of the allowed choices {choices}: {value}"
+
+    def length_error_msg(idx, length, value):
+        name = names[idx] if names else str(idx + 1)
+        return f"List argument '{name}' do not match to the previous length {length}: {value}"
+
+    def map_to_list(value, length):
+        if length is None:
+            return FakeList(value)
+        if isinstance(value, list):
+            return value
+        return [value] * length
+
+    list_len = None
+    for idx, arg in enumerate(args):
+        if isinstance(arg, list):
+            if required_types and not all(isinstance(item, required_types[idx]) for item in arg):
+                raise ConfigurationError(type_error_msg(idx, required_types[idx], arg))
+            if choices and choices[idx] and not all((item in choices[idx]) for item in arg):
+                raise ConfigurationError(value_error_msg(idx, choices[idx], arg))
+            if list_len is None:
+                list_len = len(arg)
+            elif list_len != len(arg):
+                raise ConfigurationError(length_error_msg(idx, list_len, arg))
+        else:
+            if required_types and not isinstance(arg, required_types[idx]):
+                raise ConfigurationError(type_error_msg(idx, required_types[idx], arg))
+            if choices and choices[idx] and arg not in choices[idx]:
+                raise ConfigurationError(value_error_msg(idx, choices[idx], arg))
+    if len(args) == 1:
+        return map_to_list(args[0], list_len)
+    return [map_to_list(arg, list_len) for arg in args]
+
+
+def grouper(iterable, num):
+    """Split data into fixed-length chunks"""
+    iterable = iter(iterable)
+    while True:
+        chunk = tuple(itertools.islice(iterable, num))
+        if not chunk:
+            return
+        yield chunk
 
 
 def file_open(filename, mode='r', encoding='utf8'):